my-python-backend/requirements.txt
fastapi
uvicorn
httpx
databases
sqlalchemy
asyncpg
psycopg2
pytest
pytest-asyncio
python-dotenv
